backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
httpx
PyYAML
langchain
langgraph
langchain-openai
pytest
pytest-asyncio
